Fix darwin build
