+++ pkg/leeway/build_darwin.go
@@ -1,0 +1,9 @@
+package leeway
+
+import (
+	"fmt"
+)
+
+func executeCommandsForPackageSafe(buildctx *buildContext, p *Package, wd string, commands [][]string) error {
+	return fmt.Errorf("not implemented")
+}
